% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddensity_fun.R
\name{Cgenerate}
\alias{Cgenerate}
\title{Internal function, generate matrices.}
\usage{
Cgenerate(k, p, low = -1, up = 1, kernel = "triangular")
}
\arguments{
\item{k, p}{Integer, polynomial order.}

\item{kernel}{String, the kernel function, can be \code{triangular} (default),
\code{uniform} or \code{epanechnikov}.}
}
\value{
Returns a matrix.
}
\description{
\code{Cgenerate} generates a matrix.
}
\details{
This is an internal function, and should not be called by users.
}
\keyword{internal}
