% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddensity_fun.R
\name{Cminusgenerate}
\alias{Cminusgenerate}
\title{Internal function, generate matrices.}
\usage{
Cminusgenerate(k, p, kernel = "triangular")
}
\arguments{
\item{k, p}{Integer, polynomial order.}

\item{kernel}{String, the kernel function, can be \code{triangular} (default),
\code{uniform} or \code{epanechnikov}.}
}
\value{
Returns a matrix.
}
\description{
\code{Cminusgenerate} generates a matrix.
}
\details{
This is an internal function, and should not be called by users.
}
\keyword{internal}
