% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddensity_fun.R
\name{Gminusgenerate}
\alias{Gminusgenerate}
\title{Internal function, generate matrices.}
\usage{
Gminusgenerate(p, kernel = "triangular")
}
\arguments{
\item{p}{Integer, polynomial order.}

\item{kernel}{String, the kernel function, can be \code{triangular} (default),
\code{uniform} or \code{epanechnikov}.}
}
\value{
Returns a matrix.
}
\description{
\code{Gminusgenerate} generates a matrix.
}
\details{
This is an internal function, and should not be called by users.
}
\keyword{internal}
