% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddensity_fun.R
\name{h_opt_density}
\alias{h_opt_density}
\title{Internal function, calculates theoretical optimal bandwidth.}
\usage{
h_opt_density(x, p, N, dgp_F1, dgp_Fp1, f_low, f_up, kernel = "triangular")
}
\arguments{
\item{x}{Numeric vector or data matrix, the running variable.}

\item{p}{Integer, polynomial order.}

\item{N}{Integer, sample size.}

\item{dgp_F1, dgp_Fp1}{Numeric, theoretical d.g.p.}

\item{f_low, f_up}{Numeric, lower and upper boundaries.}

\item{kernel}{String, the kernel function, can be \code{triangular} (default),
\code{uniform} or \code{epanechnikov}.}
}
\value{
Returns MSE optimal bandwidth.
}
\description{
\code{h_opt_density} calculates theoretical optimal bandwidth.
}
\details{
This is an internal function, and should not be called by users.
}
\keyword{internal}
