% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddensity_fun.R
\name{h_opt_density_res}
\alias{h_opt_density_res}
\title{Internal function, calculates theoretical optimal bandwidth.}
\usage{
h_opt_density_res(
  p,
  N,
  dgp_F1_l,
  dgp_F1_r,
  dgp_Fp1_l,
  dgp_Fp1_r,
  kernel = "triangular"
)
}
\arguments{
\item{p}{Integer, polynomial order.}

\item{N}{Integer, sample size.}

\item{kernel}{String, the kernel function, can be \code{triangular} (default),
\code{uniform} or \code{epanechnikov}.}

\item{x}{Numeric vector or data matrix, the running variable.}

\item{dgp_F1, dgp_Fp1}{Numeric, theoretical d.g.p.}

\item{f_low, f_up}{Numeric, lower and upper boundaries.}
}
\value{
Returns MSE optimal bandwidth.
}
\description{
\code{h_opt_density_res} calculates theoretical optimal bandwidth for restricted
  model.
}
\details{
This is an internal function, and should not be called by users.
}
\keyword{internal}
