% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbwdensity.R
\name{print.CJMrdbwdensity}
\alias{print.CJMrdbwdensity}
\title{Internal function.}
\usage{
\method{print}{CJMrdbwdensity}(x, ...)
}
\arguments{
\item{x}{Class \code{CJMrdbwdensity} objects.}
}
\description{
Internal function.
}
\keyword{internal}
