% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddensity.R
\name{print.CJMrddensity}
\alias{print.CJMrddensity}
\title{Internal function.}
\usage{
\method{print}{CJMrddensity}(x, ...)
}
\arguments{
\item{x}{Class \code{CJMrddensity} objects.}
}
\description{
Internal function.
}
\keyword{internal}
