% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddensity_fun.R
\name{Psigenerate}
\alias{Psigenerate}
\title{Internal function, generate matrices.}
\usage{
Psigenerate(p)
}
\arguments{
\item{p}{Integer, polynomial order.}
}
\value{
Returns a matrix.
}
\description{
\code{Psigenerate} generates a matrix.
}
\details{
This is an internal function, and should not be called by users.
}
\keyword{internal}
