% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddensity_fun.R
\name{rddensity_fV}
\alias{rddensity_fV}
\title{Internal function, implements density test.}
\usage{
rddensity_fV(
  Y,
  X,
  Nl,
  Nr,
  Nlh,
  Nrh,
  hl,
  hr,
  p,
  s,
  kernel,
  fitselect,
  vce,
  massPoints
)
}
\arguments{
\item{Y}{Numeric vector or data matrix, the estimated c.d.f.}

\item{X}{Numeric vector or data matrix, the running variable}

\item{Nl, Nr}{Integers, sample sizes to the left and right of cutoff.}

\item{Nlh, Nrh}{Integers, sample sizes to the left and right of cutoff,
within bandwidth}

\item{hl, hr}{Numeric, bandwidth to the left and right of cutoff.}

\item{p}{Integer, polynomial order}

\item{s}{Integer, higher order derivative estimate}

\item{kernel}{String, the kernel function, can be \code{triangular} (default),
\code{uniform} or \code{epanechnikov}.}

\item{fitselect}{String, the model, can be \code{restricted} or \code{unrestricted}}

\item{vce}{String, specifies the procedure used to compute the variance-covariance matrix estimator. Options are:
\code{"plugin"} for asymptotic plug-in standard errors. \code{"jackknife"} for jackknife standard errors.}

\item{massPoints}{Boolean, whether whether point estimates and standard errors}
}
\value{
Returns a data frame for further use.
}
\description{
\code{rddensity_fV} generates primitive results for the density test.
}
\details{
This is an internal function, and should not be called by users.
  NOTE: data is assumed to be on ascending order.
}
\keyword{internal}
