% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddensity_fun.R
\name{rddensity_H}
\alias{rddensity_H}
\title{Internal function, normal distribution related quantities.}
\usage{
rddensity_H(x, p)
}
\arguments{
\item{x}{Numeric, point of evaluation.}

\item{p}{Integer, polynomial order}
}
\value{
Returns density or derivatives.
}
\description{
\code{rddensity_H} calculates normal densities and derivatives.
}
\details{
This is an internal function, and should not be called by users.
}
\keyword{internal}
