% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddensity_senate.R
\docType{data}
\name{rddensity_senate}
\alias{rddensity_senate}
\alias{margin}
\title{RD Senate Data}
\format{
Numeric vector containing 1390 observations:
\describe{
  \item{margin}{Numeric vector. See Cattaneo, Frandsen and Titiunik (2015) regarding details about this dataset.}
}
}
\source{
Cattaneo, M. D., B. Frandsen, and R. Titiunik. 2015. \href{https://rdpackages.github.io/references/Cattaneo-Frandsen-Titiunik_2015_JCI.pdf}{Randomization Inference in the Regression Discontinuity Design: An Application to the Study of Party Advantages in the U.S. Senate.} \emph{Journal of Causal Inference} 3(1): 1-24. \doi{10.1515/jci-2013-0010}
}
\description{
Extract of the dataset constructed by Cattaneo, Frandsen, and Titiunik (2015),
  which include measures of incumbency advantage in the U.S. Senate for the
  period 1914-2010.
}
