% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddensity_fun.R
\name{rddensityUnique}
\alias{rddensityUnique}
\title{Internal function.}
\usage{
rddensityUnique(x)
}
\arguments{
\item{x}{Numeric vector, already sorted in ascending order.}
}
\value{
\item{unique}{A vector containing unique elements in \code{x}.}
\item{freq}{The frequency of each element in \code{x}.}
\item{index}{The last occurrence of each element in \code{x}.}
}
\description{
Find unique elements and their frequencies in a numeric vector. This function
  has a similar performance as the built-in R function \code{unique}.
}
\keyword{internal}
