% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddensity_fun.R
\name{Sgenerate}
\alias{Sgenerate}
\title{Internal function, generate matrices.}
\usage{
Sgenerate(p, low = -1, up = 1, kernel = "triangular")
}
\arguments{
\item{p}{Integer, polynomial order.}

\item{low, up}{Numeric, lower and upper limits of integration, -1 and 1 by
default.}

\item{kernel}{String, the kernel function, can be \code{triangular} (default),
\code{uniform} or \code{epanechnikov}.}
}
\value{
Returns a matrix.
}
\description{
\code{Sgenerate} generates a matrix.
}
\details{
This is an internal function, and should not be called by users.
}
\keyword{internal}
