% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbwdensity.R
\name{summary.CJMrdbwdensity}
\alias{summary.CJMrdbwdensity}
\title{Internal function.}
\usage{
\method{summary}{CJMrdbwdensity}(object, ...)
}
\arguments{
\item{object}{Class \code{CJMrdbwdensity} objects.}
}
\description{
Internal function.
}
\keyword{internal}
