% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddensity.R
\name{summary.CJMrddensity}
\alias{summary.CJMrddensity}
\title{Internal function.}
\usage{
\method{summary}{CJMrddensity}(object, ...)
}
\arguments{
\item{object}{Class \code{CJMrddensity} objects.}
}
\description{
Internal function.
}
\keyword{internal}
